Reorganizing render, deploy, destroy to unify stages input_vars, tf_objects, checks, and state_imports
Closes #1081
